ADD Classes Bilhetes

+++ codigo/Companhia_aerea/src/Ifidelidade.java
@@ -1,5 +1,5 @@
 public interface Ifidelidade {
 
-	double calcularPontos();
+	int calcularPontos();
 
 }
